requirements.txt
gensim==4.0.1
imageio==2.9.0
matplotlib==3.4.3
pandas==1.3.1
numpy==1.21.1
opencv-python==4.5.3.56
paddlepaddle-gpu==2.1.2.post101
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
tqdm==4.62.0
